+++ include/constants/file.h
@@ -74,6 +74,8 @@
 #define OVERLAY_BATTLEFORMCHANGECHECK 139
 #define OVERLAY_CHECKDEFENDERITEMEFFECTONHIT 140
 #define OVERLAY_SERVERFIELDCONDITIONCHECK 141
+#define OVERLAY_CALCDAMAGEOVERALL 142
+#define OVERLAY_CALC_CHAINOTHERMODIFIERS 143
 
 #define MAX_ACTIVE_OVERLAYS 8
