+++ overlays.mk
@@ -56,6 +56,14 @@ SERVERFIELDCONDITIONCHECK_LINK = $(BUILD)/serverfieldconditioncheck_linked.o
 SERVERFIELDCONDITIONCHECK_OUTPUT = $(BUILD)/output_serverfieldconditioncheck.bin
 OVERLAY_OUTPUTS += $(SERVERFIELDCONDITIONCHECK_OUTPUT)
 LINKED_OUTPUTS += $(SERVERFIELDCONDITIONCHECK_LINK)
+CALCDAMAGEOVERALL_LINK = $(BUILD)/calcdamageoverall_linked.o
+CALCDAMAGEOVERALL_OUTPUT = $(BUILD)/output_calcdamageoverall.bin
+OVERLAY_OUTPUTS += $(CALCDAMAGEOVERALL_OUTPUT)
+LINKED_OUTPUTS += $(CALCDAMAGEOVERALL_LINK)
+CALC_CHAINOTHERMODIFIERS_LINK = $(BUILD)/calc_chainothermodifiers_linked.o
+CALC_CHAINOTHERMODIFIERS_OUTPUT = $(BUILD)/output_calc_chainothermodifiers.bin
+OVERLAY_OUTPUTS += $(CALC_CHAINOTHERMODIFIERS_OUTPUT)
+LINKED_OUTPUTS += $(CALC_CHAINOTHERMODIFIERS_LINK)
 
 
 BATTLE_C_SRCS := $(wildcard $(C_SUBDIR)/battle/*.c)
@@ -118,6 +126,16 @@ SERVERFIELDCONDITIONCHECK_C_SRCS := $(C_SUBDIR)/individual/ServerFieldConditionC
 SERVERFIELDCONDITIONCHECK_C_OBJS := $(patsubst $(C_SUBDIR)/%.c,$(BUILD)/%.o,$(SERVERFIELDCONDITIONCHECK_C_SRCS))
 SERVERFIELDCONDITIONCHECK_OBJS := $(SERVERFIELDCONDITIONCHECK_C_OBJS) build/thumb_help.d
 OVERLAY_OBJS += $(SERVERFIELDCONDITIONCHECK_OBJS)
+
+CALCDAMAGEOVERALL_C_SRCS := $(C_SUBDIR)/individual/CalcDamageOverall.c
+CALCDAMAGEOVERALL_C_OBJS := $(patsubst $(C_SUBDIR)/%.c,$(BUILD)/%.o,$(CALCDAMAGEOVERALL_C_SRCS))
+CALCDAMAGEOVERALL_OBJS := $(CALCDAMAGEOVERALL_C_OBJS) build/thumb_help.d
+OVERLAY_OBJS += $(CALCDAMAGEOVERALL_OBJS)
+
+CALC_CHAINOTHERMODIFIERS_C_SRCS := $(C_SUBDIR)/individual/Calc_ChainOtherModifiers.c
+CALC_CHAINOTHERMODIFIERS_C_OBJS := $(patsubst $(C_SUBDIR)/%.c,$(BUILD)/%.o,$(CALC_CHAINOTHERMODIFIERS_C_SRCS))
+CALC_CHAINOTHERMODIFIERS_OBJS := $(CALC_CHAINOTHERMODIFIERS_C_OBJS) build/thumb_help.d
+OVERLAY_OBJS += $(CALC_CHAINOTHERMODIFIERS_OBJS)
 
 
 $(FIELD_LINK):$(FIELD_OBJS) rom_gen.ld
@@ -197,4 +215,16 @@ $(SERVERFIELDCONDITIONCHECK_LINK):$(SERVERFIELDCONDITIONCHECK_OBJS) rom_gen_batt
 	$(LD) $(LDFLAGS_SWITCHINABILITYCHECK) -o $@ $(SERVERFIELDCONDITIONCHECK_OBJS)
 
 $(SERVERFIELDCONDITIONCHECK_OUTPUT):$(SERVERFIELDCONDITIONCHECK_LINK)
+	$(OBJCOPY) -O binary $< $@
+
+$(CALCDAMAGEOVERALL_LINK):$(CALCDAMAGEOVERALL_OBJS) rom_gen_battle.ld
+	$(LD) $(LDFLAGS_SWITCHINABILITYCHECK) -o $@ $(CALCDAMAGEOVERALL_OBJS)
+
+$(CALCDAMAGEOVERALL_OUTPUT):$(CALCDAMAGEOVERALL_LINK)
+	$(OBJCOPY) -O binary $< $@
+
+$(CALC_CHAINOTHERMODIFIERS_LINK):$(CALC_CHAINOTHERMODIFIERS_OBJS) rom_gen_battle.ld
+	$(LD) $(LDFLAGS_SWITCHINABILITYCHECK) -o $@ $(CALC_CHAINOTHERMODIFIERS_OBJS)
+
+$(CALC_CHAINOTHERMODIFIERS_OUTPUT):$(CALC_CHAINOTHERMODIFIERS_LINK)
 	$(OBJCOPY) -O binary $< $@
