+++ scripts/make.py
@@ -53,6 +53,8 @@ OUTPUT_CALCBASEDAMAGE = 'build/output_calcbasedamage.bin'
 OUTPUT_BATTLEFORMCHANGECHECK = 'build/output_battleformchangecheck.bin'
 OUTPUT_CHECKDEFENDERITEMEFFECTONHIT = 'build/output_checkdefenderitemeffectonhit.bin'
 OUTPUT_SERVERFIELDCONDITIONCHECK = 'build/output_serverfieldconditioncheck.bin'
+OUTPUT_CALCDAMAGEOVERALL = 'build/output_calcdamageoverall.bin'
+OUTPUT_CALC_CHAINOTHERMODIFIERS = 'build/output_calc_chainothermodifiers.bin'
 BYTE_REPLACEMENT = 'bytereplacement'
 HOOKS = 'hooks'
 ARM_HOOKS = 'armhooks'
@@ -60,7 +62,7 @@ REPOINTS = 'repoints'
 ROUTINE_POINTERS = 'routinepointers'
 
 NEW_OVERLAYS_NO_ARM9_EXT = [OUTPUT_BATTLE, OUTPUT_FIELD, OUTPUT_POKEDEX, OUTPUT_GETMONEVOLUTION, OUTPUT_GETMONEVOLUTION_BATTLE, OUTPUT_MOVEHITDEFENDERABILITYCHECK, OUTPUT_SWITCHINABILITYCHECK, OUTPUT_STATBUFFCHANGE,
-                            OUTPUT_CALCBASEDAMAGE, OUTPUT_BATTLEFORMCHANGECHECK, OUTPUT_CHECKDEFENDERITEMEFFECTONHIT, OUTPUT_SERVERFIELDCONDITIONCHECK]
+                            OUTPUT_CALCBASEDAMAGE, OUTPUT_BATTLEFORMCHANGECHECK, OUTPUT_CHECKDEFENDERITEMEFFECTONHIT, OUTPUT_SERVERFIELDCONDITIONCHECK, OUTPUT_CALCDAMAGEOVERALL, OUTPUT_CALC_CHAINOTHERMODIFIERS]
 
 LINKED_SECTIONS = ['build/linked.o', 'build/battle_linked.o', 'build/field_linked.o', 'build/pokedex_linked.o']
 OFFSET_START_IN_129 = 0x600
